requirements.txt
fastmcp
python-dotenv
requests
urllib3
